# src/main/java/com/freemmy/reducing/routine/rest/gen/api/CartsApi.java
/*** NOTE: This class is auto generated by the swagger code generator program (3.3.4).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.freemmy.reducing.routine.rest.gen.api;

import com.freemmy.reducing.routine.rest.gen.model.CartDTO;
import com.freemmy.reducing.routine.rest.gen.model.ErrorDTO;
import com.freemmy.reducing.routine.rest.gen.model.ItemDTO;
import com.freemmy.reducing.routine.rest.gen.model.NewItemDTO;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

import javax.validation.Valid;

@Api(value = "Carts", description = "the Carts API")
public interface CartsApi {

    CartsApiDelegate getDelegate();

    @ApiOperation(value = "Add item to cart", nickname = "addItem", notes = "", response = ItemDTO.class, tags = {
            "carts", })

    @ApiResponses(value = {
            @ApiResponse(code = 201, message = "Item is added to the cart", response = ItemDTO.class),
            @ApiResponse(code = 200, message = "unexpected error", response = ErrorDTO.class) })
    @RequestMapping(value = "/carts/{cart_id}/items",
            produces = { "application/json" },
            consumes = { "application/json" },
            method = RequestMethod.POST)
    default ResponseEntity<ItemDTO> addItem(
            @ApiParam(value = "ID of the cart.", required = true) @PathVariable("cart_id") Long cartId,
            @ApiParam(value = "", required = true) @Valid @RequestBody NewItemDTO newItemDTO) {
        return getDelegate().addItem(cartId, newItemDTO);
    }

    @ApiOperation(value = "Get cart", nickname = "getCart", notes = "", response = CartDTO.class, tags = { "carts", })

    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Cart", response = CartDTO.class),
            @ApiResponse(code = 200, message = "unexpected error", response = ErrorDTO.class) })
    @RequestMapping(value = "/carts/{cart_id}",
            produces = { "application/json" },
            method = RequestMethod.GET)
    default ResponseEntity<CartDTO
            > getCart(
            @ApiParam(value = "ID of the cart.", required = true) @PathVariable("cart_id") Long cartId) {
        return getDelegate().getCart(cartId);
    }

}